The physical back button is now working correctly and does not return you to the register page.
Also, fixed a navigation issue where after unregistering the device in the settings menu and being returned to the register page, the back button would lead back to the settings page and then the home page. Now the back button will only exit the app when used on the register page.